split long-time unittests to smaller ones
